+++ integration_tests/test_bergermeer.py
@@ -25,7 +25,6 @@ def test_integration(tmp_path):
         unittests_data_path / "v2_bergermeer.sqlite",
         unittests_data_path / "dem_test_5m.tif",
         tmp_path / "gridadmin.h5",
-        model_area_path=None,  # untested
         meta={
             "model_slug": "slug-123abc",
             "revision_hash": "123abc",